implemented expand macros tests and function
